fix(privacymodal): moved privacymodal to organism
