+++ source/components/organisms/PrivacyModal/PrivacyModal.types.ts
